CR: Move authenticated state to a non-shared file
