+++ ui/admin/tests/e2e/tests/credential-store-static.spec.js
@@ -7,7 +7,8 @@ import { test, expect } from '@playwright/test';
 import { readFile } from 'fs/promises';
 import { nanoid } from 'nanoid';
 
-import { checkEnv, authenticatedState } from '../helpers/general';
+import { authenticatedState } from '../global-setup';
+import { checkEnv } from '../helpers/general';
 import {
   authenticateBoundaryCli,
   authorizeSessionByTargetIdCli,
