+++ ui/admin/tests/e2e/tests/delete-resources.spec.js
@@ -6,7 +6,8 @@
 import { test } from '@playwright/test';
 import { execSync } from 'node:child_process';
 
-import { checkEnv, authenticatedState } from '../helpers/general';
+import { authenticatedState } from '../global-setup';
+import { checkEnv } from '../helpers/general';
 import {
   authenticateBoundaryCli,
   checkBoundaryCli,
