+++ ui/admin/tests/e2e/tests/dynamic-host-catalog-aws.spec.js
@@ -6,7 +6,8 @@
 import { test, expect } from '@playwright/test';
 import { nanoid } from 'nanoid';
 
-import { checkEnv, authenticatedState } from '../helpers/general';
+import { authenticatedState } from '../global-setup';
+import { checkEnv } from '../helpers/general';
 import {
   authenticateBoundaryCli,
   deleteScopeCli,
