+++ ui/admin/tests/e2e/tests/group-role.spec.js
@@ -5,7 +5,7 @@
 
 import { test } from '@playwright/test';
 
-import { authenticatedState } from '../helpers/general';
+import { authenticatedState } from '../global-setup';
 import {
   authenticateBoundaryCli,
   checkBoundaryCli,
