+++ ui/admin/tests/e2e/tests/session-recording-aws-ent.spec.js
@@ -6,7 +6,8 @@
 import { test, expect } from '@playwright/test';
 import { execSync } from 'child_process';
 
-import { checkEnv, authenticatedState } from '../helpers/general';
+import { authenticatedState } from '../global-setup';
+import { checkEnv } from '../helpers/general';
 import {
   authenticateBoundaryCli,
   checkBoundaryCli,
